Move otelfunc to middleware/ namespace
The package is a Middleware implementation, not part of the core SDK.

Sibling layout matches triggers/blob and leaves room for future middleware

(retry policies, auth, request logging, etc.) without further migration.
